Illuminate first, then mask
Previously the mask making the transmission target was created first
then it was pointwise multiplied by the speckle signal.

An improvement has resulted in terms of execution speed: on my pc
it passed from 9s to 6s, because multiplication is now restricted
only to the region where the three slits are located which amount
to a third of the whole signal domain.